+++ pkg/gardenadm/botanist/operatingsystemconfig.go
@@ -12,7 +12,6 @@ import (
 	"time"
 
 	systemddbus "github.com/coreos/go-systemd/v22/dbus"
-	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/tools/record"
@@ -227,87 +226,4 @@ func (b *AutonomousBotanist) IsGardenerNodeAgentInitialized(ctx context.Context)
 	}
 
 	return !exists, nil
-}
-
-// FakeOSC is a dummy implementation of operatingsystemconfig.Interface.
-// It creates and returns a static user data secret. This is just a temporary implementation for testing the Worker
-// deployment until we have a real OperatingSystemConfig for `gardenadm bootstrap`.
-// TODO(timebertt): replace this with a proper OperatingSystemConfig component implementation
-type FakeOSC struct {
-	Client client.Client
-
-	ControlPlaneWorkerPool, ControlPlaneNamespace string
-}
-
-const fakeUserDataSecretName = "user-data"
-
-// Deploy implements operatingsystemconfig.Interface.
-func (f *FakeOSC) Deploy(ctx context.Context) error {
-	userDataSecret := &corev1.Secret{
-		TypeMeta: metav1.TypeMeta{
-			APIVersion: corev1.SchemeGroupVersion.String(),
-			Kind:       "Secret",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      fakeUserDataSecretName,
-			Namespace: f.ControlPlaneNamespace,
-		},
-		Data: map[string][]byte{
-			extensionsv1alpha1.OperatingSystemConfigSecretDataKey: []byte("#!/usr/bin/env bash\necho Hello gardenadm!"),
-		},
-	}
-
-	return f.Client.Patch(ctx, userDataSecret, client.Apply, client.FieldOwner("gardenadm"), client.ForceOwnership)
-}
-
-// WorkerPoolNameToOperatingSystemConfigsMap implements operatingsystemconfig.Interface.
-func (f *FakeOSC) WorkerPoolNameToOperatingSystemConfigsMap() map[string]*operatingsystemconfig.OperatingSystemConfigs {
-	return map[string]*operatingsystemconfig.OperatingSystemConfigs{
-		f.ControlPlaneWorkerPool: {
-			Init: operatingsystemconfig.Data{
-				SecretName: ptr.To(fakeUserDataSecretName),
-			},
-		},
-	}
-}
-
-// Destroy implements operatingsystemconfig.Interface.
-func (*FakeOSC) Destroy(context.Context) error { panic("implement me") }
-
-// Restore implements operatingsystemconfig.Interface.
-func (*FakeOSC) Restore(context.Context, *gardencorev1beta1.ShootState) error { panic("implement me") }
-
-// Migrate implements operatingsystemconfig.Interface.
-func (*FakeOSC) Migrate(context.Context) error { panic("implement me") }
-
-// WaitMigrate implements operatingsystemconfig.Interface.
-func (*FakeOSC) WaitMigrate(context.Context) error { panic("implement me") }
-
-// Wait implements operatingsystemconfig.Interface.
-func (*FakeOSC) Wait(context.Context) error { panic("implement me") }
-
-// WaitCleanup implements operatingsystemconfig.Interface.
-func (*FakeOSC) WaitCleanup(context.Context) error { panic("implement me") }
-
-// DeleteStaleResources implements operatingsystemconfig.Interface.
-func (*FakeOSC) DeleteStaleResources(context.Context) error { panic("implement me") }
-
-// WaitCleanupStaleResources implements operatingsystemconfig.Interface.
-func (*FakeOSC) WaitCleanupStaleResources(context.Context) error { panic("implement me") }
-
-// SetAPIServerURL implements operatingsystemconfig.Interface.
-func (*FakeOSC) SetAPIServerURL(string) { panic("implement me") }
-
-// SetCABundle implements operatingsystemconfig.Interface.
-func (*FakeOSC) SetCABundle(string) { panic("implement me") }
-
-// SetSSHPublicKeys implements operatingsystemconfig.Interface.
-func (*FakeOSC) SetSSHPublicKeys([]string) { panic("implement me") }
-
-// SetClusterDNSAddresses implements operatingsystemconfig.Interface.
-func (*FakeOSC) SetClusterDNSAddresses([]string) { panic("implement me") }
-
-// SetCredentialsRotationStatus implements operatingsystemconfig.Interface.
-func (*FakeOSC) SetCredentialsRotationStatus(*gardencorev1beta1.ShootCredentialsRotation) {
-	panic("implement me")
 }
