+++ pkg/gardenlet/operation/botanist/worker.go
@@ -29,7 +29,8 @@ import (
 // DefaultWorker creates the default deployer for the Worker custom resource.
 func (b *Botanist) DefaultWorker() worker.Interface {
 	workers := b.Shoot.GetInfo().Spec.Provider.Workers
-	// in `gardenadm bootstrap` we only deploy the control plane worker pool
+	// In `gardenadm bootstrap` we only deploy the control plane worker pool. When running `gardenadm init` on the
+	// medium-touch control plane, the full `Worker` with all pools will be deployed.
 	if b.Shoot.IsAutonomous() && !b.Shoot.RunsControlPlane() {
 		workers = []gardencorev1beta1.Worker{*v1beta1helper.ControlPlaneWorkerPoolForShoot(b.Shoot.GetInfo().DeepCopy())}
 	}
